requirements.txt
redis>=4.5.5
python-dotenv>=1.0.0